Extract aggregations into separate module
